[Enhancement #303] Load and display term snapshot detail.
