+++ src/action/__tests__/AsyncTermActions.test.ts
@@ -7,18 +7,20 @@ import { ThunkDispatch } from "../../util/Types";
 import {
   createTerm,
   loadDefinitionRelatedTermsTargeting,
+  loadTerm,
   setTermDefinitionSource,
   setTermStatus,
 } from "../AsyncTermActions";
 import TermOccurrence from "../../model/TermOccurrence";
 import Generator from "../../__tests__/environment/Generator";
 import Term, { CONTEXT as TERM_CONTEXT } from "../../model/Term";
-import ActionType from "../ActionType";
+import ActionType, { AsyncActionSuccess, MessageAction } from "../ActionType";
 import MessageType from "../../model/MessageType";
 import { langString } from "../../model/MultilingualString";
 import Constants from "../../util/Constants";
 import AsyncActionStatus from "../AsyncActionStatus";
 import TermStatus from "../../model/TermStatus";
+import { verifyExpectedAssets } from "../../__tests__/environment/TestUtil";
 
 jest.mock("../../util/Routing");
 jest.mock("../../util/Ajax", () => {
@@ -353,6 +355,118 @@ describe("AsyncTermActions", () => {
           );
         expect(action).toBeDefined();
         expect(action.payload).toEqual(TermStatus.CONFIRMED);
+      });
+    });
+  });
+
+  describe("load vocabulary term", () => {
+    it("loads vocabulary term using term and vocabulary normalized names on call", () => {
+      store.getState().user = Generator.generateUser();
+      const vocabName = "test-vocabulary";
+      const termName = "test-term";
+      Ajax.get = jest
+        .fn()
+        .mockImplementation(() =>
+          Promise.resolve(require("../../rest-mock/terms")[0])
+        );
+      return Promise.resolve(
+        (store.dispatch as ThunkDispatch)(
+          loadTerm(termName, { fragment: vocabName })
+        )
+      ).then(() => {
+        const url = (Ajax.get as jest.Mock).mock.calls[0][0];
+        expect(url).toEqual(
+          Constants.API_PREFIX +
+            "/vocabularies/" +
+            vocabName +
+            "/terms/" +
+            termName
+        );
+      });
+    });
+
+    it("passes namespace parameter when it is specified on action call", () => {
+      store.getState().user = Generator.generateUser();
+      const vocabName = "test-vocabulary";
+      const termName = "test-term";
+      const namespace =
+        "http://onto.fel.cvut.cz/ontologies/termit/vocabularies/";
+      Ajax.get = jest
+        .fn()
+        .mockImplementation(() =>
+          Promise.resolve(require("../../rest-mock/terms")[0])
+        );
+      return Promise.resolve(
+        (store.dispatch as ThunkDispatch)(
+          loadTerm(termName, {
+            fragment: vocabName,
+            namespace,
+          })
+        )
+      ).then(() => {
+        const url = (Ajax.get as jest.Mock).mock.calls[0][0];
+        expect(url).toEqual(
+          Constants.API_PREFIX +
+            "/vocabularies/" +
+            vocabName +
+            "/terms/" +
+            termName
+        );
+        const config = (Ajax.get as jest.Mock).mock.calls[0][1];
+        expect(config).toBeDefined();
+        expect(config.getParams().namespace).toEqual(namespace);
+      });
+    });
+
+    it("uses public API endpoint to fetch single vocabulary term when user is not authenticated", () => {
+      const term = require("../../rest-mock/terms")[0];
+      Ajax.get = jest.fn().mockImplementation(() => Promise.resolve(term));
+      return Promise.resolve(
+        (store.dispatch as ThunkDispatch)(
+          loadTerm("test-term", { fragment: "test-vocabulary" })
+        )
+      ).then((data: AsyncActionSuccess<Term> | MessageAction) => {
+        const url = (Ajax.get as jest.Mock).mock.calls[0][0];
+        expect(url).toContain(Constants.PUBLIC_API_PREFIX);
+        verifyExpectedAssets(
+          [term],
+          [(data as AsyncActionSuccess<Term>).payload]
+        );
+      });
+    });
+
+    it("uses version-based endpoint when snapshot timestamp is provided", () => {
+      store.getState().user = Generator.generateUser();
+      const vocabName = "test-vocabulary";
+      const termName = "test-term";
+      const namespace =
+        "http://onto.fel.cvut.cz/ontologies/termit/vocabularies/";
+      const timestamp = "20220731T100000Z";
+      Ajax.get = jest
+        .fn()
+        .mockImplementation(() =>
+          Promise.resolve(require("../../rest-mock/terms")[0])
+        );
+      return Promise.resolve(
+        (store.dispatch as ThunkDispatch)(
+          loadTerm(
+            termName,
+            {
+              fragment: vocabName,
+              namespace,
+            },
+            timestamp
+          )
+        )
+      ).then(() => {
+        const url = (Ajax.get as jest.Mock).mock.calls[0][0];
+        expect(url).toEqual(
+          `${Constants.API_PREFIX}/vocabularies/${vocabName}/terms/${termName}/versions`
+        );
+        const config = (Ajax.get as jest.Mock).mock.calls[0][1];
+        expect(config).toBeDefined();
+        expect(config.getParams().namespace).toEqual(namespace);
+        expect(config.getParams().at).toEqual(timestamp);
       });
     });
   });
