+++ src/component/term/TermDetail.tsx
@@ -4,7 +4,6 @@ import { RouteComponentProps, withRouter } from "react-router";
 import { connect } from "react-redux";
 import { ThunkDispatch } from "../../util/Types";
 import {
-  loadTerm,
   loadVocabulary,
   removeTerm,
   updateTerm,
@@ -40,6 +39,7 @@ import { DefinitionRelatedChanges } from "./DefinitionRelatedTermsEdit";
 import TermOccurrence from "../../model/TermOccurrence";
 import {
   approveOccurrence,
+  loadTerm,
   removeOccurrence,
 } from "../../action/AsyncTermActions";
 import TermReadOnlyIcon from "./authorization/TermReadOnlyIcon";
@@ -50,7 +50,7 @@ export interface CommonTermDetailProps extends HasI18n {
   term: Term | null;
   vocabulary: Vocabulary;
   loadVocabulary: (iri: IRI) => void;
-  loadTerm: (termName: string, vocabularyIri: IRI) => void;
+  loadTerm: (termName: string, vocabularyIri: IRI, timestamp?: string) => void;
 }
 
 interface TermDetailProps
@@ -108,11 +108,16 @@ export class TermDetail extends EditableComponent<
   private loadTerm(): void {
     const vocabularyName: string = this.props.match.params.name;
     const termName: string = this.props.match.params.termName;
+    const timestamp: string | undefined = this.props.match.params.timestamp;
     const namespace = Utils.extractQueryParam(
       this.props.location.search,
       "namespace"
     );
-    this.props.loadTerm(termName, { fragment: vocabularyName, namespace });
+    this.props.loadTerm(
+      termName,
+      { fragment: vocabularyName, namespace },
+      timestamp
+    );
   }
 
   public componentDidUpdate(prevProps: TermDetailProps) {
@@ -297,8 +302,8 @@ export default connect(
   (dispatch: ThunkDispatch) => {
     return {
       loadVocabulary: (iri: IRI) => dispatch(loadVocabulary(iri)),
-      loadTerm: (termName: string, vocabularyIri: IRI) =>
-        dispatch(loadTerm(termName, vocabularyIri)),
+      loadTerm: (termName: string, vocabularyIri: IRI, timestamp?: string) =>
+        dispatch(loadTerm(termName, vocabularyIri, timestamp)),
       updateTerm: (term: Term) => dispatch(updateTerm(term)),
       removeTerm: (term: Term) => dispatch(removeTerm(term)),
       approveOccurrence: (occurrence: TermOccurrence) =>
